Clean up
* Removed controllers and views not necessary
* Moved previews into mockup controller
* Classes and views now have their final names
