+++ module/Application/config/module.config.php
@@ -35,16 +35,6 @@ return [
                     ],
                 ],
             ],
-            'application/admin' => [
-                'type'    => Segment::class,
-                'options' => [
-                    'route'    => '/application/admin[/:action]',
-                    'defaults' => [
-                        'controller' => Controller\AdminController::class,
-                        'action'     => 'index',
-                    ],
-                ],
-            ],
         ],
     ],
     'controllers' => [
@@ -61,7 +51,6 @@ return [
         'exception_template'       => 'error/index',
         'template_map' => [
             'layout/layout'           => __DIR__ . '/../view/layout/layout.phtml',
-            'layout/content-only'     => __DIR__ . '/../view/layout/content-only.phtml',
             'error/404'               => __DIR__ . '/../view/error/404.phtml',
             'error/index'             => __DIR__ . '/../view/error/index.phtml',
         ],
