--- module/Application/view/application/admin/index.phtml
@@ -1,5 +1,0 @@
-<div class="row">
-    <div class="col-lg-12">
-        <h1>Admin home</h1>
-    </div>
-</div>
