+++ module/Content/config/module.config.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace Content;
 
 use Dlayer\ViewHelper;
+use Zend\Mvc\Controller\ControllerManager;
 use Zend\Router\Http\Literal;
 use Zend\Router\Http\Segment;
 use Zend\ServiceManager\Factory\InvokableFactory;
@@ -28,12 +29,22 @@ return [
                 ],
             ],
             'content/manager' => [
-                'type'    => Segment::class,
+                'type'    => Literal::class,
                 'options' => [
-                    'route'    => '/content/manager[/:action]',
+                    'route'    => '/content/manager',
                     'defaults' => [
                         'controller' => Controller\ManagerController::class,
                         'action'     => 'index',
+                    ],
+                ],
+            ],
+            'content/mockup' => [
+                'type'    => Segment::class,
+                'options' => [
+                    'route'    => '/content/mockup[/:action]',
+                    'defaults' => [
+                        'controller' => Controller\MockupController::class,
+                        'action'     => 'row-selected',
                     ],
                 ],
             ],
@@ -42,16 +53,15 @@ return [
     'controllers' => [
         'factories' => [
             Controller\IndexController::class => InvokableFactory::class,
-            Controller\ManagerController::class => InvokableFactory::class
+            Controller\ManagerController::class => InvokableFactory::class,
+            Controller\MockupController::class => InvokableFactory::class
         ],
     ],
     'view_helpers' => [
         'factories' => [
-            View\Helper\TestHelper::class => InvokableFactory::class,
             ViewHelper\Toolbar::class => InvokableFactory::class
         ],
         'aliases' => [
-            'testHelper' => View\Helper\TestHelper::class,
             'toolbar' => ViewHelper\Toolbar::class
         ]
     ],
@@ -66,7 +76,7 @@ return [
             'layout/content-manager-column-selected'     => __DIR__ . '/../view/layout/content-manager-column-selected.phtml',
             'layout/content-manager-tool-selected'     => __DIR__ . '/../view/layout/content-manager-tool-selected.phtml',
             'layout/content-manager-tool-and-item-selected'     => __DIR__ . '/../view/layout/content-manager-tool-and-item-selected.phtml',
-            'layout/content-manager-mock-tools'     => __DIR__ . '/../view/layout/content-manager-mock-tools.phtml',
+            'layout/content-manager-designer'     => __DIR__ . '/../view/layout/designer.phtml',
         ],
         'template_path_stack' => [
             __DIR__ . '/../view',
