+++ module/Content/view/content/index/index.phtml
@@ -1,20 +1,22 @@
 <div class="row">
     <div class="col-xl-12">
-        <h2>Content manager development</h2>
+        <h2>Content Manager</h2>
+
+        <p class="lead">Content Manager development and previews.</p>
+
+        <h3>Dev</h3>
+
+        <ul>
+            <li><a href="/content/manager">Manager</a> - Content manager development</li>
+        </ul>
 
         <h3>Static previews of Content manager state</h3>
 
         <ul>
-            <li><a href="/content/manager/row-selected">Row selected</a> - A row has been selected</li>
-            <li><a href="/content/manager/column-selected">Column selected</a> - A column has been selected</li>
-            <li><a href="/content/manager/tool-selected">Tool selected</a> - Add a content item/row to a column</li>
-            <li><a href="/content/manager/tool-and-item-selected">Tool and item selected</a> - Edit the selected content item</li>
-        </ul>
-
-        <h3>Mocking tools development</h3>
-
-        <ul>
-            <li><a href="/content/manager/mock-tools">Mock</a> - Mocking up tools development</li>
+            <li><a href="/content/mockup/row-selected">Row selected</a> - A row has been selected</li>
+            <li><a href="/content/mockup/column-selected">Column selected</a> - A column has been selected</li>
+            <li><a href="/content/mockup/tool-selected">Tool selected</a> - Add a content item/row to a column</li>
+            <li><a href="/content/mockup/tool-and-item-selected">Tool and item selected</a> - Edit the selected content item</li>
         </ul>
     </div>
 </div>
