+++ module/Content/view/content/mockup/column-selected.phtml
