+++ module/Content/view/content/mockup/tool-selected.phtml
